+++ src/core/mst-operations.ts
@@ -379,12 +379,18 @@ export function detach<T extends IMSTNode>(thing: T): T {
  */
 export function destroy(thing: IMSTNode) {
     const node = getMSTAdministration(thing)
-    node.detach()
-    node.die()
+    if (node.isRoot)
+        node.die()
+    else
+        node.parent!.removeChild(node.subpath)
 }
 
 export function isAlive(thing: IMSTNode): boolean {
     return getMSTAdministration(thing).isAlive
+}
+
+export function addDisposer(thing: IMSTNode, disposer: () => void) {
+    getMSTAdministration(thing).addDisposer(disposer)
 }
 
 export function getEnv(thing: IMSTNode): any {