+++ src/types/complex-types/object.ts
@@ -16,7 +16,7 @@ import {
     hasOwnProperty,
     isPlainObject
 } from "../../utils"
-import { MSTAdministration, maybeMST, getType, IMSTNode, getMSTAdministration, getSnapshot,IJsonPatch } from "../../core"
+import { MSTAdministration, maybeMST, getType, IMSTNode, getMSTAdministration, getSnapshot, IJsonPatch } from "../../core"
 import { IType, IComplexType, isType } from "../type"
 import { ComplexType } from "./complex-type"
 import { getPrimitiveFactoryFromValue } from "../primitives"
@@ -51,7 +51,7 @@ export class ObjectType extends ComplexType<any, any> {
 
     constructor(name: string, baseModel: Object) {
         super(name)
-        Object.seal(baseModel) // make sure nobody messes with it
+        Object.freeze(baseModel) // make sure nobody messes with it
         this.baseModel = baseModel
         invariant(/^\w[\w\d_]*$/.test(name), `Typename should be a valid identifier: ${name}`)
         this.modelConstructor = new Function(`return function ${name} (){}`)() // fancy trick to get a named function...., http://stackoverflow.com/questions/5905492/dynamic-function-name-in-javascript